backend-makers/requirements.txt
# FastAPI y servidor
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Configuración y validación
python-dotenv==1.0.0
pydantic==2.4.2

# Base de datos - Supabase
supabase==2.3.0

# LLM - Google Gemini
google-generativeai==0.3.2

# Utilidades
requests==2.31.0
httpx==0.25.0 